feat: Add a General Triage view.
Fixes #264
This view is for the rotating triage of Firefox::General and
Toolkit::General. It closely reflects the official query on bugdash.
Unfortunately it can't notify you when it's your turn to triage, but
it's a start. These bugs will also feature a new Severity column. Since
there's a new Severity column, we now explain the severity levels in the
priority guide. Because part of this involves tracking stale needinfos,
I also added a tag for these bugs that shows who the needinfo is for.
And because these queries are pretty complex, I added a tooltip to the
bug list titles that explains the query in more detail. Under the hood,
we now request release and build info from other Mozilla APIs, and
provide that to the whole component in GlobalContext.releases. So this
can be used in the future for other purposes. I also made the server
cache code generic, since we now have three separate server caches.