style: move the puzzle icon postfix to bug id.
when a bug has an OMC Jira ticket but isn't synced with the ticket, we
show a puzzle icon next to it with a tooltip explaining the situation.
before this commit, the puzzle icon was shown next to the ticket. but
since it's really describing the bug more than the ticket, this commit
moves it so it's next to the bug id. also, this fixes an issue where the
severity column header was not capitalized, and adds the puzzle icon to
bugs in the Jira Tickets view, if applicable.